[fixmystreet.com] Survey banner
Big purple survey banner, which is hidden by default, and shown on 40%
of pageloads, by JavaScript.

Works on both the static pages, and the map pages.

Clicking the CTA button, or the close icon, hides the banner and sets
a cookie so that the banner isn’t shown again on that device.

Banner is always hidden during the mobile reporting flow.

Fixes #2928.